aaaa

+++ server/server.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const morgan = require('morgan');
 const { json, download } = require('express/lib/response');
-const { param, body, validationResult } = require('express-validator');
+const { param, body, validationResult, check } = require('express-validator');
 const passport = require('passport');
 const LocalStrategy = require('passport-local').Strategy;
 const session = require('express-session');
@@ -52,7 +52,7 @@ const isLoggedIn = (req, res, next) => {
     if (req.isAuthenticated())
         return next();
 
-    return res.status(401).json({ error: 'not authenticated' });
+    return res.status(401).json({ error: 'Not authenticated' });
 }
 
 app.use(cors(corsOptions));
@@ -61,6 +61,65 @@ app.use(express.json());
 app.use(morgan('dev'));
 app.use(passport.initialize());
 app.use(passport.session());
+
+//Validation function
+async function checkCode(value) {
+    const codes = await courseDAO.getAllCoursesCode();
+    if (codes.find(c => c === value)) {
+        return true;
+    }
+    return false;
+}
+
+function PlanValidation() {
+
+    return [
+        body('available')
+            .exists().withMessage('Plan type must be present')
+            .isInt([0, 1]).withMessage("Plan type not available."),
+        body('plan.*.codice')
+            .exists().withMessage("Codice must be preset.")
+            .isString().withMessage("Codice must be a string.")
+            .isLength({ min: 7, max: 7 }).withMessage("Codice length not valid.")
+            .custom(async (value) => {
+                const check = await checkCode(value);
+                if (!check) {
+                    throw new Error('Code doenst match with any courses');
+                }
+                return true;
+            })
+        ,
+        body('plan.*.tot_studenti')
+            .optional({ nullable: true })
+            .isInt()
+            .withMessage('Tot studenti invalid value.'),
+        body('plan.*.max_studenti')
+            .optional({ nullable: true })
+            .isInt().withMessage('Max studenti invalid value.'),
+        body('plan.*.crediti')
+            .exists().withMessage('Crediti must be present.')
+            .isInt().withMessage('Crediti invalid value.'),
+        body('plan')
+            .isArray().withMessage('Plan invalid format')
+            .custom((value, { req }) => {
+                const max_crediti = req.body.available ? { min: 60, max: 80 } : { min: 20, max: 40 };
+                const planCredits = value.map(p => p.crediti).reduce((prev, next) => prev + next, 0);
+                if (planCredits > max_crediti.max || planCredits < max_crediti.min) {
+                    throw new Error('Number of credit out of bound.');
+                }
+                return true;
+            }),
+        body('plan.*')
+            .optional()
+            .if(val => val.max_studenti != null)
+            .custom((value) => {
+                if (value.tot_studenti >= value.max_studenti) {
+                    throw new Error('Too many students.');
+                }
+                return true;
+            })
+    ]
+}
 
 
 //courses API
@@ -73,7 +132,14 @@ app.get('/api/courses', (req, res) => {
 });
 
 //user API
-app.post('/api/sessions', function (req, res, next) {
+app.post('/api/sessions', [
+    body('username')
+        .isEmail().withMessage('Email not valid')
+], function (req, res, next) {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).json({ message: errors.array()[0].msg });
+    }
     passport.authenticate('local', (err, user, info) => {
         if (err) return next(err);
         if (!user) return res.status(401).json(info);
@@ -99,22 +165,24 @@ app.get('/api/sessions/current', (req, res) => {
 
 //plan API
 app.get('/api/plans', isLoggedIn, (req, res) => {
-     // TODO: if req.user.available => validation
+    // TODO: if req.user.available => validation
     planDAO.getPlanByUser(req.user.id)
         .then(plan => res.json(plan))
         .catch(err => res.status(500).json(err));
 })
 
-app.post('/api/plans/addPlan', isLoggedIn, async (req, res) => {
-
-    //TODO: validation and if errors then return status code 422 (Unprocessable entity)
+app.post('/api/plans/addPlan', isLoggedIn, PlanValidation(), async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).json({ errors: errors.array() });
+    }
     try {
         await Promise.all(req.body.plan.map(course => planDAO.addCourseToPlan(req.user.id, course.codice))); //array of proms that execute the query for the single course
         await userDAO.updateType(req.body.available, req.user.id);
         await courseDAO.updateStudentsCount();
-        res.status(200).json({type: req.body.available, plan: req.body.plan});
+        res.status(200).json({ type: req.body.available, plan: req.body.plan });
     } catch (error) {
-        res.status(503).json({errMessage: `Database error during the creation of plan for user ${req.user.id}.`});
+        res.status(503).json({ errMessage: `Database error during the creation of plan for user ${req.user.id}.` });
         return;
     }
 });
@@ -123,10 +191,14 @@ app.delete('/api/plans/deletePlan', isLoggedIn, (req, res) => {
     // TODO: if req.user.available => validation
     planDAO.deletePlan(req.user.id).then(() => courseDAO.updateStudentsCount()).then(() => userDAO.updateType(null, req.user.id))
         .then(() => res.status(200).json({ message: "Delete success" }))
-        .catch(err => res.status(503).send({ errMessage: `Delete went wrong for the user ${req.user.id}`}));
+        .catch(() => res.status(503).send({ errMessage: `Delete went wrong for the user ${req.user.id}` }));
 });
 
-app.put('/api/plans/updatePlan', isLoggedIn, async (req, res) => {
+app.put('/api/plans/updatePlan', isLoggedIn, PlanValidation(), async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).json({ errors: errors.array() });
+    }
     try {
         await planDAO.deletePlan(req.user.id);
         await Promise.all(req.body.plan.map(course => planDAO.addCourseToPlan(req.user.id, course.codice)));
@@ -134,7 +206,7 @@ app.put('/api/plans/updatePlan', isLoggedIn, async (req, res) => {
         await courseDAO.updateStudentsCount();
         res.status(200).json(req.body);
     } catch (error) {
-        res.status(503).json({ errMessage: `Plan update failed for user ${req.user.id}`})
+        res.status(503).json({ errMessage: `Plan update failed for user ${req.user.id}` })
     }
 
 });
